+++ airgeddon.sh
@@ -60,7 +60,6 @@ optional_tools_names=(
 						"reaver"
 						"bully"
 						"pixiewps"
-						"unbuffer"
 						"bettercap"
 						"beef"
 						"packetforge-ng"
@@ -100,14 +99,13 @@ declare -A possible_package_names=(
 									[${optional_tools_names[14]}]="reaver" #reaver
 									[${optional_tools_names[15]}]="bully" #bully
 									[${optional_tools_names[16]}]="pixiewps" #pixiewps
-									[${optional_tools_names[17]}]="expect / expect-dev" #unbuffer
-									[${optional_tools_names[18]}]="bettercap" #bettercap
-									[${optional_tools_names[19]}]="beef-xss / beef-project" #beef
-									[${optional_tools_names[20]}]="aircrack-ng" #packetforge-ng
-									[${optional_tools_names[21]}]="hostapd-wpe" #hostapd-wpe
-									[${optional_tools_names[22]}]="asleap" #asleap
-									[${optional_tools_names[23]}]="john" #john
-									[${optional_tools_names[24]}]="openssl" #openssl
+									[${optional_tools_names[17]}]="bettercap" #bettercap
+									[${optional_tools_names[18]}]="beef-xss / beef-project" #beef
+									[${optional_tools_names[19]}]="aircrack-ng" #packetforge-ng
+									[${optional_tools_names[20]}]="hostapd-wpe" #hostapd-wpe
+									[${optional_tools_names[21]}]="asleap" #asleap
+									[${optional_tools_names[22]}]="john" #john
+									[${optional_tools_names[23]}]="openssl" #openssl
 									[${update_tools[0]}]="curl" #curl
 								)
 
@@ -5012,16 +5010,16 @@ function initialize_menu_options_dependencies() {
 	et_captive_portal_dependencies=("${optional_tools_names[5]}" "${optional_tools_names[6]}" "${optional_tools_names[7]}" "${optional_tools_names[11]}" "${optional_tools_names[12]}")
 	wash_scan_dependencies=("${optional_tools_names[13]}")
 	reaver_attacks_dependencies=("${optional_tools_names[14]}")
-	bully_attacks_dependencies=("${optional_tools_names[15]}" "${optional_tools_names[17]}")
-	bully_pixie_dust_attack_dependencies=("${optional_tools_names[15]}" "${optional_tools_names[16]}" "${optional_tools_names[17]}")
+	bully_attacks_dependencies=("${optional_tools_names[15]}")
+	bully_pixie_dust_attack_dependencies=("${optional_tools_names[15]}" "${optional_tools_names[16]}")
 	reaver_pixie_dust_attack_dependencies=("${optional_tools_names[14]}" "${optional_tools_names[16]}")
-	et_sniffing_sslstrip2_dependencies=("${optional_tools_names[5]}" "${optional_tools_names[6]}" "${optional_tools_names[7]}" "${optional_tools_names[18]}" "${optional_tools_names[19]}")
-	wep_attack_dependencies=("${optional_tools_names[2]}" "${optional_tools_names[20]}")
-	enterprise_attack_dependencies=("${optional_tools_names[21]}" "${optional_tools_names[22]}" "${optional_tools_names[24]}")
-	asleap_attacks_dependencies=("${optional_tools_names[22]}")
-	john_attacks_dependencies=("${optional_tools_names[23]}")
-	johncrunch_attacks_dependencies=("${optional_tools_names[23]}" "${optional_tools_names[1]}")
-	enterprise_certificates_dependencies=("${optional_tools_names[24]}")
+	et_sniffing_sslstrip2_dependencies=("${optional_tools_names[5]}" "${optional_tools_names[6]}" "${optional_tools_names[7]}" "${optional_tools_names[17]}" "${optional_tools_names[18]}")
+	wep_attack_dependencies=("${optional_tools_names[2]}" "${optional_tools_names[19]}")
+	enterprise_attack_dependencies=("${optional_tools_names[20]}" "${optional_tools_names[21]}" "${optional_tools_names[23]}")
+	asleap_attacks_dependencies=("${optional_tools_names[21]}")
+	john_attacks_dependencies=("${optional_tools_names[22]}")
+	johncrunch_attacks_dependencies=("${optional_tools_names[22]}" "${optional_tools_names[1]}")
+	enterprise_certificates_dependencies=("${optional_tools_names[23]}")
 }
 
 #Set possible changes for some commands that can be found in different ways depending of the O.S.
@@ -5778,8 +5776,8 @@ function beef_pre_menu() {
 	language_strings "${language}" 266
 	print_simple_separator
 
-	if [[ "${beef_found}" -eq 0 ]] && [[ ${optional_tools[${optional_tools_names[19]}]} -eq 1 ]]; then
-		if [[ ${optional_tools[${optional_tools_names[5]}]} -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[6]}]} -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[7]}]} -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[18]}]} -eq 1 ]]; then
+	if [[ "${beef_found}" -eq 0 ]] && [[ ${optional_tools[${optional_tools_names[18]}]} -eq 1 ]]; then
+		if [[ ${optional_tools[${optional_tools_names[5]}]} -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[6]}]} -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[7]}]} -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[17]}]} -eq 1 ]]; then
 			language_strings "${language}" 409 "warning"
 			language_strings "${language}" 416 "pink"
 		else
@@ -5821,7 +5819,7 @@ function beef_pre_menu() {
 			fi
 		;;
 		2)
-			if [[ "${beef_found}" -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[19]}]} -eq 1 ]]; then
+			if [[ "${beef_found}" -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[18]}]} -eq 1 ]]; then
 				echo
 				language_strings "${language}" 412 "red"
 				language_strings "${language}" 115 "read"
@@ -8921,7 +8919,7 @@ function set_wps_attack_script() {
 			;;
 		esac
 	else
-		unbuffer="unbuffer "
+		unbuffer="stdbuf -i0 -o0 -e0 "
 		case ${wps_attack_mode} in
 			"pindb"|"custompin")
 				attack_cmd1="bully \${script_interface} -b \${script_wps_bssid} -c \${script_wps_channel} \${script_bully_reaver_band_modifier} -L -F -B -v ${bully_verbosity} -p "
@@ -10324,7 +10322,7 @@ function kill_beef() {
 	debug_print
 
 	local beef_pid
-	beef_pid="$(ps -C "${optional_tools_names[19]}" --no-headers -o pid | tr -d ' ')"
+	beef_pid="$(ps -C "${optional_tools_names[18]}" --no-headers -o pid | tr -d ' ')"
 	if ! kill "${beef_pid}" &> /dev/null; then
 		if ! kill "$(ps -C "beef" --no-headers -o pid | tr -d ' ')" &> /dev/null; then
 			kill "$(ps -C "ruby" --no-headers -o pid,cmd | grep "beef" | awk '{print $1}')" &> /dev/null
@@ -10369,7 +10367,7 @@ function prepare_beef_start() {
 	debug_print
 
 	valid_possible_beef_path=0
-	if [[ ${beef_found} -eq 0 ]] && [[ ${optional_tools[${optional_tools_names[19]}]} -eq 0 ]]; then
+	if [[ ${beef_found} -eq 0 ]] && [[ ${optional_tools[${optional_tools_names[18]}]} -eq 0 ]]; then
 		language_strings "${language}" 405 "blue"
 		ask_yesno 191 "yes"
 		if [ "${yesno}" = "y" ]; then
@@ -10386,12 +10384,12 @@ function prepare_beef_start() {
 			language_strings "${language}" 413 "yellow"
 			language_strings "${language}" 115 "read"
 		fi
-	elif [[ "${beef_found}" -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[19]}]} -eq 0 ]]; then
+	elif [[ "${beef_found}" -eq 1 ]] && [[ ${optional_tools[${optional_tools_names[18]}]} -eq 0 ]]; then
 		fix_beef_executable "${beef_path}"
 		echo
 		language_strings "${language}" 413 "yellow"
 		language_strings "${language}" 115 "read"
-	elif [[ "${beef_found}" -eq 0 ]] && [[ ${optional_tools[${optional_tools_names[19]}]} -eq 1 ]]; then
+	elif [[ "${beef_found}" -eq 0 ]] && [[ ${optional_tools[${optional_tools_names[18]}]} -eq 1 ]]; then
 		language_strings "${language}" 405 "blue"
 		ask_yesno 415 "yes"
 		if [ "${yesno}" = "y" ]; then
@@ -10457,7 +10455,7 @@ function fix_beef_executable() {
 	echo -e "./beef"
 	} >> "/usr/bin/beef"
 	chmod +x "/usr/bin/beef" > /dev/null 2>&1
-	optional_tools[${optional_tools_names[19]}]=1
+	optional_tools[${optional_tools_names[18]}]=1
 
 	rewrite_script_with_custom_beef "set" "${1}"
 }
@@ -10485,8 +10483,8 @@ function start_beef_service() {
 
 	debug_print
 
-	if ! service "${optional_tools_names[19]}" restart > /dev/null 2>&1; then
-		systemctl restart "${optional_tools_names[19]}.service" > /dev/null 2>&1
+	if ! service "${optional_tools_names[18]}" restart > /dev/null 2>&1; then
+		systemctl restart "${optional_tools_names[18]}.service" > /dev/null 2>&1
 	fi
 }
 
@@ -10514,9 +10512,9 @@ function launch_beef() {
 			global_process_pid=""
 		fi
 	else
-		manage_output "-hold -bg \"#000000\" -fg \"#00FF00\" -geometry ${g4_middleright_window} -T \"BeEF\"" "${optional_tools_names[19]}" "BeEF"
+		manage_output "-hold -bg \"#000000\" -fg \"#00FF00\" -geometry ${g4_middleright_window} -T \"BeEF\"" "${optional_tools_names[18]}" "BeEF"
 		if [ "${AIRGEDDON_WINDOWS_HANDLING}" = "tmux" ]; then
-			get_tmux_process_id "{optional_tools_names[19]}"
+			get_tmux_process_id "{optional_tools_names[18]}"
 			et_processes+=("${global_process_pid}")
 			global_process_pid=""
 		fi